laba three

+++ lesson32/app/src/main/java/com/example/lesson3/Fragment1.kt
@@ -24,7 +24,7 @@ class Fragment1 : Fragment() {
     private fun nextButtonClick(){
         binding.next.setOnClickListener{
             val activityCollback = requireActivity() as ActivityCollback
-            Calculator.num1 = binding.editText1.text.toString().toInt()
+            Calculator.num1 = binding.editText1.text.toString().toDouble()
             activityCollback.showNextFragment(x)
         }
     }
